Move WorkflowRun to separate module
